pyproject.toml
[project]
name = "slack-proxy"
version = "0.1.0"
description = "Slack channel proxy with SQLite persistence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.0",
  "fastapi>=0.116.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.0",
  "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.uv]
package = false
